Defining GraphQL Schema and Resolvers
