+++ package.json
@@ -1,6 +1,14 @@
 {
+  "name": "streamverse",
+  "version": "1.0.0",
+  "description": "A video streaming application",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server/server.js"
+  },
   "dependencies": {
     "apollo-server-express": "^3.12.0",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
@@ -10,10 +18,5 @@
     "mongoose": "^7.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
-  },
-  
-  "scripts": {
-    "start": "node server.js"
   }
-  
 }
