+++ server/server.js
@@ -2,20 +2,19 @@ const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
 const jwt = require('jsonwebtoken');
 
-// Import your resolvers and type definitions
-const typeDefs = require('./graphql/schema');
+const typeDefs = require('./graphql/schema/schema');
 const resolvers = require('./graphql/resolvers');
 
 // get JWT secret from environment variables
 const SECRET = process.env.JWT_SECRET;
 
 const authMiddleware = (req, res, next) => {
-    const token = req.headers.authorization;
-    if (token) {
-        const user = jwt.verify(token, SECRET);
-        req.user = user;
-    }
-    next();
+  const token = req.headers.authorization;
+  if (token) {
+    const user = jwt.verify(token, SECRET);
+    req.user = user;
+  }
+  next();
 };
 
 const app = express();
@@ -23,19 +22,21 @@ const app = express();
 // use middleware in your Express server
 app.use(authMiddleware);
 
-// Create an Apollo server and apply it to the Express app
-const server = new ApolloServer({ 
-    typeDefs, 
-    resolvers, 
-    context: ({ req }) => ({ user: req.user }) 
-});
+async function startServer() {
+  const server = new ApolloServer({
+    typeDefs,
+    resolvers,
+    context: ({ req }) => ({ user: req.user }),
+  });
 
-server.applyMiddleware({ app });
+  await server.start();
 
-// Define a port
-const PORT = process.env.PORT || 3001;
+  server.applyMiddleware({ app });
 
-// Start the server
-app.listen(PORT, () => {
+  const PORT = process.env.PORT || 3001;
+  app.listen(PORT, () => {
     console.log(`Server is running at http://localhost:${PORT}${server.graphqlPath}`);
-});
+  });
+}
+
+startServer();
